+++ Sources/BartyCrouchKit/Globals/Secret.swift
@@ -2,5 +2,5 @@ import Foundation
 
 public enum Secret: Equatable {
     case microsoftTranslator(secret: String)
-    case deepl(secret: String)
+    case deepL(secret: String)
 }