+++ Tests/BartyCrouchTranslatorTests/DeepLTranslatorApiTests.swift
@@ -3,21 +3,21 @@ import Foundation
 import Microya
 import XCTest
 
-class DeeplTranslatorApiTests: XCTestCase {
+class DeepLTranslatorApiTests: XCTestCase {
     func testTranslate() {
         let apiKey = "" // TODO: load from environment variable
         guard !apiKey.isEmpty else { return }
 
-        let endpoint = DeeplApi.translate(
+        let endpoint = DeepLApi.translate(
             texts: ["How old are you?", "Love"],
             from: .english,
             to: .german,
             apiKey: apiKey
         )
 
-        let apiProvider = ApiProvider<DeeplApi>()
+        let apiProvider = ApiProvider<DeepLApi>()
 
-        switch apiProvider.performRequestAndWait(on: endpoint, decodeBodyTo: DeeplTranslateResponse.self) {
+        switch apiProvider.performRequestAndWait(on: endpoint, decodeBodyTo: DeepLTranslateResponse.self) {
         case let .success(translateResponses):
             XCTAssertEqual(translateResponses.translations[0].text, "Wie alt sind Sie?")